AnotherParser

Documentation for AnotherParser.

AnotherParser allows one to implement a recursive descent parser given a hierarchical grammar expressed as BNF.

Grammars

A grammar is represented by a BNFGrammar struct. AllGrammars is a catalog of all of the defined grammars.

AnotherParser.BNFGrammar — Type

BNFGrammar

Represents a single grammar which can consist of a number of DerivationRules.

AnotherParser.AllGrammars — Constant

A Dict of all defined grammars.

Each BNFGrammar includes a collection of [DerivationRule]@ref)s. Each DerivationRule has a "left hand side' that is a tree of BNFNodes.

A DerivationRule can be referred to by name via BNFRef, and thus shared among different parts of the grammar and between grammars.

AnotherParser.Alternatives — Type

Alternatives(nodes...)

Matches any one element of nodes.

AnotherParser.BNFNode — Type

BNFNode

Abstract supertype for all structs that we use to implement a grammar.

AnotherParser.BNFRef — Type

BNFRef(grammar, name)

delegates to the "left hand side" of the DerivationRule named name in grammar.

AnotherParser.CharacterLiteral — Type

CharacterLiteral(c)

Matches the single character c.

AnotherParser.Constructor — Type

Constructor(node, constructor_function)

Apply constructor_function to the result of recognizing node and return that as the result.

AnotherParser.DerivationRule — Type

DerivationRule(grammar, rule_name, expression)

Implements a single production named name in the specified grammar. One can include expression in other expressions using BNFRef(grammar, rule_name). The rule can have a constructor function, which, if present, will be called with the recognized value, and the context.

AnotherParser.Empty — Type

Empty()

Succeed while matching nothing.

AnotherParser.EndOfInput — Type

EndOfInput()

Succeed if input is exhausted.

AnotherParser.RegexNode — Type

RegexNode <: BNFNode(re::Regex)

Match the specified regular expression." So that the parser can access captures, The second return value of recognize is the RegexMatch object returned by match.

AnotherParser.Repeat — Type

Repeat(n::BNFNode; min=0, max=typemax(Int))

Matches repeated occurances of n. The minimum and maxmum number of allowed matches can be specified.

AnotherParser.Sequence — Type

Sequence(nodes...)

Successively match each of nodes in turn.

AnotherParser.StringCollector — Type

StringCollector

StringCollector returns the entire substrring of the input that @trace trace_recognize was recognized by its subexpression.

AnotherParser.StringLiteral — Type

StringLiteral(str)

Matches the string str.

Each BNFNode implements the recognize generic function, which performs the actual parsing:

AnotherParser.recognize — Function

recognize(::BNFNode, input::AbstractString; index, finish, context)

Attempt to parse input as the specified BNFNode, starting at index. Return three values: whether the node matched the input, the parsed value represented by the matched input, and the next index into input. finish is the index into input of the last character to be considered. The context argument is passed to constructor functions (see Constructor and DerivationRule) but is otherwise unused.

Utility Functions

AnotherParser.@bnf_str — Macro

bnf"str"grammar_name

Parse str as BNF and add those productions to the grammar named grammar_name.

AnotherParser.show_grammar — Function

show_grammar(grammar_name::Symbol)

Show a pretty description of all of the derivations in the grammar.

AnotherParser.check_references — Function

check_references(grammar)::Bool

Warn of any BNFRef nodes in the grammar that don't have a target. Return true if there are any issues.

AnotherParser.pretty — Function

pretty(::BNFNode)

Return a human-readable string that describes the node.

AnotherParser.is_left_recursive — Function

is_left_recursive(node::BNFNode, grammar::Symbol, name::AbstractString)

Returns true is node is left recursive with respect to the DerivationRule named name.

AnotherParser.walk_nodes — Function

walk_nodes(f, node)

Applies f to each node in the node tree that descends from node.

Construction Utilities

A grammar has little use recognizing its input if it does not also build a useful data structure from the abstract syntax tree.

All BNFNode types should describe what their recognize methods return as a value. For CharacterLiteral and StringLiteral it is just the matched character or string. For RegexNode it is the regular expression's match object. For Alternatives it is the value of whatever subexpression matched. For Sequence it is a Vector of the matched values of the subexpressions.

The Constructor BNFNode type can be used to wrap another BNFNode expression to massage the value that it returns.

Also, DerivationRule can be assigned a constructor property.

Various functions are provided to serve as constructors:

AnotherParser.flatten_to_string — Function

flatten_to_string(v)

Flatten a tail recursive Vector of characters to a string.

Constructor Functions

The application usually wants to construct some form of data structure during the parsing operation, not just a simple tree representing the result of the parse. Both Constructor and DerivationRule allow for a constructor function to be called on the currently recognized input and return instead an object that is meaningful to the application. Each constructor function takes as arguments:

a context object that is meaningless to everything but the constructor function;

the entire input string;

the start index into that input string string of the portion currently recognized;

the index of the first character in the input after the recognized item;

the subordinate fragments that were just parsed.

Example

using AnotherParser

# A grammar is named by a symbol
BNFGrammar(:example)

# "text" can be one word, or a sequence of words separated by spaces:
# Note the two uses of Constructor so that the result will be a
# single, flat Tuple of words:
DerivationRule(
    :example,
    "text",
    Alternatives(
        Constructor(BNFRef(:example, "word"),
                    (context, input::AbstractString, from::Int, to::Int, x) -> (x,)),
        Constructor(
            Sequence(BNFRef(:example, "word"),
                     BNFRef(:example, "space"),
                     BNFRef(:example, "text")),
            (context, input::AbstractString, from::Int, to::Int, x) -> (x[1], x[3]...))))

# Collapse the successive characters of a word into a string:
DerivationRule(
    :example,
    "word",
    StringCollector(BNFRef(:example, "word1")))

# "word1" is a sequence of letters:
DerivationRule(
    :example,
    "word1",
    Alternatives(BNFRef(:example, "letter"),
                 Sequence(BNFRef(:example, "letter"),
                          BNFRef(:example, "word1"))))

# Lowercase letters
DerivationRule(
    :example,
    "letter",
    Alternatives([CharacterLiteral(c) for c in 'a':'z']...))

# Whitespace:
DerivationRule(
    :example,
    "space",
    Alternatives(
        CharacterLiteral(' '),
        Sequence(
            CharacterLiteral(' '),
            BNFRef(:example, "space"))))

# recognize will return a tuple of "word"s and
# the next index of the input string
recognize(AllGrammars[:example]["text"],
          "this  is    a test")

(true, ("this", "is", "a", "test"), 19)

There is much room for simplification and syntactic sugar.

Until I write more documentation, see the example grammars in the examples directory.

Debugging

AnotherParser.debug_parsing — Function

debug_parsing(grammar::BNFGrammar, rulename::AbstractString, input::AbstractString; index = 1, finish = length(input), context=nothing, report_file::AbstractString)

Run the parser specified by grammar and rulename to parse input from index to finish with the specified context object. The parsing log is written to report_file as HTML. That file presents a hierarchichal view of how the parse progressed.

AnotherParser.DEBUG_BNFNODES — Constant

DEBUG_BNFNODES

DEBUG_BNFNODES is a list of the node uid's of nodes that should be @info logged when they match the current input at the specified index.

Index

AnotherParser.AllGrammars

AnotherParser.DEBUG_BNFNODES

AnotherParser.Alternatives

AnotherParser.BNFGrammar

AnotherParser.BNFNode

AnotherParser.BNFRef

AnotherParser.CharacterLiteral

AnotherParser.Constructor

AnotherParser.DerivationRule

AnotherParser.Empty

AnotherParser.EndOfInput

AnotherParser.RegexNode

AnotherParser.Repeat

AnotherParser.Sequence

AnotherParser.StringCollector

AnotherParser.StringLiteral

AnotherParser.check_references

AnotherParser.debug_parsing

AnotherParser.flatten_to_string

AnotherParser.is_left_recursive

AnotherParser.pretty

AnotherParser.recognize

AnotherParser.show_grammar

AnotherParser.walk_nodes

AnotherParser.@bnf_str