AnotherParser

Documentation for AnotherParser.

AnotherParser allows one to implement a recursive descent parser given a hierarchical grammar expressed as BNF.

ISO EBNF is not yet supported.

Grammars

A grammar is represented by a BNFGrammar struct. AllGrammars is a catalog of all of the defined grammars.

AnotherParser.BNFGrammar — Type

BNFGrammar

Represents a single grammar which can consist of a number of DerivationRules.

AnotherParser.AllGrammars — Constant

A Dict of all defined grammars.

Each BNFGrammar includes a collection of [DerivationRule]@ref)s. Each DerivationRule has a "left hand side' that is a tree of BNFNodes.

A DerivationRule can be referred to by name via BNFRef, and thus shared among different parts of the grammar and between grammars.

AnotherParser.Alternatives — Type

Alternatives(nodes...)

Matches any one element of nodes.

AnotherParser.BNFNode — Type

BNFNode

Abstract supertype for all structs that we use to implement a grammar.

AnotherParser.BNFRef — Type

BNFRef(grammar, name)

delegates to the "left hand side" of the DerivationRule named name in grammar.

AnotherParser.CharacterInSet — Type

CharacterInSet(set)

Matches any single character in set.

AnotherParser.CharacterLiteral — Type

CharacterLiteral(c)

Matches the single character c.

AnotherParser.CharacterSatisfiesPredicate — Type

CharacterSatisfiesPredicate(predicate::Function)

Matches any single character that predicate returns true for.

AnotherParser.Constructor — Type

Constructor(node, constructor_function)

Apply constructor_function to the result of recognizing node and return that as the result.

AnotherParser.DerivationRule — Type

DerivationRule(grammar, rule_name, expression)

Implements a single production named name in the specified grammar. One can include expression in other expressions using BNFRef(grammar, rule_name). The rule can have a constructor function, which, if present, will be called with the recognized value, and the context.

AnotherParser.Empty — Type

Empty()

Succeed while matching nothing.

AnotherParser.EndOfInput — Type

EndOfInput()

Succeed if input is exhausted.

AnotherParser.Excluding — Type

Excluding(exclude, match)

Match if the current input matches match, but ofly if it does not match exclude.

AnotherParser.RegexNode — Type

RegexNode <: BNFNode(re::Regex)

Match the specified regular expression." So that the parser can access captures, The second return value of recognize is the RegexMatch object returned by match.

AnotherParser.Repeat — Type

Repeat(n::BNFNode; min=0, max=typemax(Int))

Matches repeated occurances of n. The minimum and maxmum number of allowed matches can be specified.

AnotherParser.Sequence — Type

Sequence(nodes...)

Successively match each of nodes in turn.

AnotherParser.StringLiteral — Type

StringLiteral(str)

Matches the string str.

Each BNFNode implements the recognize generic function, which performs the actual parsing:

AnotherParser.recognize — Function

recognize(::BNFNode, input::AbstractString; index, finish, context)

Attempt to parse input as the specified BNFNode, starting at index. Return three values: whether the node matched the input, the parsed value represented by the matched input, and the next index into input. finish is the index into input of the last character to be considered. The context argument is passed to constructor functions (see Constructor and DerivationRule) but is otherwise unused.

Utility Functions

AnotherParser.@bnf_str — Macro

bnf"str"grammar_name

Parse str as BNF and add those productions to the grammar named grammar_name.

Since bnf_str is used to define the BNF grammar itself, and it's convenient to have the BNF strings near the hand coded BNF grammar, there's a bootstraping mode which defers parsing of str until a BNF grammar is available.

AnotherParser.show_grammar — Function

show_grammar(grammar_name::Symbol)

Show a pretty description of all of the derivations in the grammar.

AnotherParser.check_references — Function

check_references(grammar)::Bool

Warn of any BNFRef nodes in the grammar that don't have a target. Return true if there are any issues.

Returns true if any warnings were issued.

AnotherParser.pretty — Function

pretty(::BNFNode)

Return a human-readable string that describes the node.

AnotherParser.is_left_recursive — Function

is_left_recursive(::BNFGrammar)
is_left_recursive(::DerivationRule)
is_left_recursive(node::BNFNode, grammar::Symbol, name::AbstractString)

Returns true is node is left recursive with respect to the DerivationRule named name.

AnotherParser.walk_nodes — Function

walk_nodes(f, node)

Applies f to each node in the node tree that descends from node.

AnotherParser.deep_flatten — Function

deep_flatten(collection)

Flattens an arbitrarily deeply nested collection.

Construction Utilities

A grammar has little use recognizing its input if it does not also build a useful data structure from the abstract syntax tree.

All BNFNode types should describe what their recognize methods return as a value. For CharacterLiteral and StringLiteral it is just the matched character or string. For RegexNode it is the regular expression's match object. For Alternatives it is the value of whatever subexpression matched. For Sequence it is a Vector of the matched values of the subexpressions.

The Constructor BNFNode type can be used to wrap another BNFNode expression to massage the value that it returns.

Also, DerivationRule can be assigned a constructor property.

Various functions are provided to serve as constructors:

Constructor Functions

The application usually wants to construct some form of data structure during the parsing operation, not just a simple tree representing the result of the parse. Both Constructor and DerivationRule allow for a constructor function to be called on the currently recognized input and return instead an object that is meaningful to the application. Each constructor function takes as arguments:

a context object that is meaningless to everything but the constructor function;

the entire input string;

the start index into that input string string of the portion currently recognized;

the index of the first character in the input after the recognized item;

the subordinate fragments that were just parsed.

AnotherParser.identity_constructor_function — Function

identity_constructor_function(context, input::AbstractString, from::Int, to::Int, value)

A constructor function that just returns the raw value from its associated DerivationRule or subexpression.

AnotherParser.substring_constructor_function — Function

substring_constructor_function(context, input::AbstractString, from::Int, to::Int, value)

A constructor function that ruturns the matched substring.

AnotherParser.value_is_from_index — Function

value_is_from_index(context, input::AbstractString, from::Int, to::Int, value)

Uses the from index as the constructor value for the associated node. This is useful if some higher BNFNode will be using SubString to construct its value.

AnotherParser.value_is_to_index — Function

value_is_to_index(context, input::AbstractString, from::Int, to::Int, value) = to

Uses the to index as the constructor value for the associated node. This is useful if some higher BNFNode will be using SubString to construct its value.

AnotherParser.value_is — Function

value_is(value)

Returns a constructor function that always returns value.

AnotherParser.flattening_constructor_function — Function

flattening_constructor_function(context, input::AbstractString, from::Int, to::Int, value)

Returns a flattened list given an arbitrarily nested one.

Example

using AnotherParser

# A grammar is named by a symbol
BNFGrammar(:example)

# "text" can be one word, or a sequence of words separated by spaces:
# Note the two uses of Constructor so that the result will be a
# single, flat Tuple of words:
DerivationRule(
    :example,
    "text",
    Alternatives(
        Constructor(BNFRef(:example, "word"),
                    (context, input::AbstractString, from::Int, to::Int, x) -> (x,)),
        Constructor(
            Sequence(BNFRef(:example, "word"),
                     BNFRef(:example, "space"),
                     BNFRef(:example, "text")),
            (context, input::AbstractString, from::Int, to::Int, x) -> (x[1], x[3]...))))

# Collapse the successive characters of a word into a string:
DerivationRule(
    :example,
    "word",
    BNFRef(:example, "word1")).constructor = substring_constructor_function

# "word1" is a sequence of letters:
DerivationRule(
    :example,
    "word1",
    Alternatives(BNFRef(:example, "letter"),
                 Sequence(BNFRef(:example, "letter"),
                          BNFRef(:example, "word1"))))

# Lowercase letters
DerivationRule(
    :example,
    "letter",
    Alternatives([CharacterLiteral(c) for c in 'a':'z']...))

# Whitespace:
DerivationRule(
    :example,
    "space",
    Alternatives(
        CharacterLiteral(' '),
        Sequence(
            CharacterLiteral(' '),
            BNFRef(:example, "space"))))

# recognize will return a tuple of "word"s and
# the next index of the input string
recognize(AllGrammars[:example]["text"],
          "this  is    a test")

(true, ("this", "is", "a", "test"), 19)

There is much room for simplification and syntactic sugar.

Until I write more documentation, see the example grammars in the examples directory.

Additional examples are available in the ./examples directory. This includes an a grammar for parsing SemVer semantic version number format and a grammar for BNF as specified on the Wikipedia BNF page.

Debugging

AnotherParser.debug_parsing — Function

debug_parsing(grammar::BNFGrammar, rulename::AbstractString, input::AbstractString; index = 1, finish = length(input), context=nothing, report_file::AbstractString, enable_debug_logging_for = should_enable_debug_logging_for)

Run the parser specified by grammar and rulename to parse input from index to finish with the specified context object. The parsing log is written to report_file as HTML. That file presents a hierarchichal view of how the parse progressed.

enabledebuglogging_for is a function that takes a BNFNode as argument and returns true if that node should be tracked in the info log and false otherwise.

AnotherParser.DEBUG_BNFNODES — Constant

DEBUG_BNFNODES

DEBUG_BNFNODES is a list of the node uid's of nodes that should be @info logged when they match the current input at the specified index.

AnotherParser.should_enable_debug_logging_for — Function

should_enable_debug_logging_for(node::BNFNode)

The default value for the enable_debug_logging_for keyword argument to debug_parsing.

It avoids logging of CharacterLiteral nodes and of Alternatives nodes where every alternative is a `CharacterLiteral node.

The Grammar for BNF

The grammar for BNF itself is specified as BNF. The file ./examples/BNF/bnf.jl implements two BNF grammars. :BootstrapBNFGrammar is a grammar that is hand coded using BNFNode types. :BNF is automatically built by the :BootstrapBNFGrammar grammar.

There is some machinery to orchestrate that. I won't decribe it here. I hope the implementation is clear enough.

Index

AnotherParser.AllGrammars

AnotherParser.DEBUG_BNFNODES

AnotherParser.Alternatives

AnotherParser.BNFGrammar

AnotherParser.BNFNode

AnotherParser.BNFRef

AnotherParser.CharacterInSet

AnotherParser.CharacterLiteral

AnotherParser.CharacterSatisfiesPredicate

AnotherParser.Constructor

AnotherParser.DerivationRule

AnotherParser.Empty

AnotherParser.EndOfInput

AnotherParser.Excluding

AnotherParser.RegexNode

AnotherParser.Repeat

AnotherParser.Sequence

AnotherParser.StringLiteral

AnotherParser.check_references

AnotherParser.debug_parsing

AnotherParser.deep_flatten

AnotherParser.flattening_constructor_function

AnotherParser.identity_constructor_function

AnotherParser.is_left_recursive

AnotherParser.pretty

AnotherParser.recognize

AnotherParser.should_enable_debug_logging_for

AnotherParser.show_grammar

AnotherParser.substring_constructor_function

AnotherParser.value_is

AnotherParser.value_is_from_index

AnotherParser.value_is_to_index

AnotherParser.walk_nodes

AnotherParser.@bnf_str