AnotherParser.jl

AnotherParser allows one to implement a parser given a grammar expressed as BNF.

AnotherParser does not yet directly support a BNF grammar expressed in BNF syntax.

Grammars

A grammar is implemented as a tree of structs that are subtypes of BNFNode.

AnotherParser.Alternatives — Type

Alternatives(nodes...)

Matches any one element of nodes.

AnotherParser.BNFNode — Type

BNFNode

Abstract supertype for all structs that we use to implement a grammar.

AnotherParser.BNFRef — Type

BNFRef(grammar, name) delegates to the "left hand side" of the DerivationRule named name in grammar.

AnotherParser.CharacterLiteral — Type

CharacterLiteral(c)

Matches the single character c.

AnotherParser.Constructor — Type

Constructor(node, constructor_function)

Apply constructor_function to rhe result of recognizing node and return that as the result.

AnotherParser.DerivationRule — Type

DerivationRule(grammar, rule_name, expression)

Implements a single production named name in the specified grammar. One can include expression in other expressions using BNFRef(grammar, rule_name).

AnotherParser.Sequence — Type

Sequence(nodes...)

Successively match each of nodes in turn.

AnotherParser.StringCollector — Type

StringCollector StringCollector returns the entire substrring of the input that was recognized by its subexpression.

Each BNFNode implements the recognize generic function which performs the actual parsing:

AnotherParser.recognize — Method

recognize(::BNFNode, input::String; index, finish)

Attempt to parse input as the specified BNFNode. Return two values: the value represented by the matched input, and the next index into input. If the returned index is equal to the initial index then the input did not matchthe BNFNode.

AnotherParser.AllGrammars — Constant

A Dict of all defined grammars.

Example

using AnotherParser

# A grammar is named by a symbol
BNFGrammar(:example)

# "text" can be one word, or a sequence of words separated by spaces:
# Note the two uses of Constructor so that the result will be a
# single, flat Tuple of words:
DerivationRule(
    :example,
    "text",
    Alternatives(
        Constructor(BNFRef(:example, "word"),
                    x -> (x,)),
        Constructor(
            Sequence(BNFRef(:example, "word"),
                     BNFRef(:example, "space"),
                     BNFRef(:example, "text")),
            x -> (x[1], x[3]...))))

# Collapse the successive characters of a word into a string:
DerivationRule(
    :example,
    "word",
    StringCollector(BNFRef(:example, "word1")))

# "word1" is a sequence of letters:
DerivationRule(
    :example,
    "word1",
    Alternatives(BNFRef(:example, "letter"),
                 Sequence(BNFRef(:example, "letter"),
                          BNFRef(:example, "word1"))))

# Lowercase letters
DerivationRule(
    :example,
    "letter",
    Alternatives([CharacterLiteral(c) for c in 'a':'z']...))

# Whitespace:
DerivationRule(
    :example,
    "space",
    Alternatives(
        CharacterLiteral(' '),
        Sequence(
            CharacterLiteral(' '),
            BNFRef(:example, "space"))))

# recognize will return a tuple of "word"s and
# the next index of the input string
recognize(AllGrammars[:example]["text"],
          "this  is    a test")

(("this", "is", "a", "test"), 19)

There is much room for simplification and syntactic sugar.

Until I write more documentation, see test/SemVerBNF.jl for an example that implements the SemVer version number format.